Refactor devices to folders
