+++ aioshelly/block_device/__init__.py
@@ -1,0 +1,6 @@
+"""Shelly Gen1 CoAP block based device."""
+
+from .coap import COAP
+from .device import BLOCK_VALUE_UNIT, Block, BlockDevice
+
+__all__ = ["COAP", "BLOCK_VALUE_UNIT", "Block", "BlockDevice"]
