+++ aioshelly/rpc_device/__init__.py
@@ -1,0 +1,6 @@
+"""Shelly Gen2 RPC based device."""
+
+from .device import RpcDevice, UpdateType
+from .wsrpc import WsServer
+
+__all__ = ["RpcDevice", "UpdateType", "WsServer"]
